+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.1.4
+ - `addCallback` method added to both `ApiRequest` and `ApiResponse` classes
+ - `getLastRequestInfo` method added to `ApiRequest`, returns url, http method, post\put body and http headers
+
 # 1.1.3
  - Method for setting orgId in request
  - Updated method for reporting search terms